docs/requirements.txt
alabaster==0.7.12
Babel==2.9.1
bottle==0.12.19
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
conan==1.36.0
deprecation==2.0.7
distro==1.5.0
docutils==0.16
fasteners==0.16
future==0.18.2
idna==2.10
imagesize==1.2.0
importlib-metadata==4.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
node-semver==0.6.1
packaging==20.9
patch-ng==1.17.4
pluginbase==1.0.0
pyenchant==3.2.0
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
Sphinx==4.0.1
sphinx-argparse==0.2.5
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-spelling==7.2.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.25.11
zipp==3.4.1
